requirements.txt
about-time==4.2.1
aiohttp==3.8.6
aiosignal==1.3.1
alive-progress==3.1.4
appnope==0.1.3
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.1.0
backcall==0.2.0
bitarray==2.8.2
certifi==2023.7.22
charset-normalizer==3.3.1
comm==0.1.4
cytoolz==0.12.2
debugpy==1.8.0
decorator==5.1.1
eth-abi==4.2.1
eth-account==0.9.0
eth-hash==0.5.2
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.5.1
eth-utils==2.3.0
executing==2.0.0
frozenlist==1.4.0
grapheme==0.6.0
hexbytes==0.3.1
idna==3.4
ipykernel==6.26.0
ipython==8.16.1
jedi==0.19.1
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
jupyter_client==8.5.0
jupyter_core==5.4.0
lru-dict==1.2.0
matplotlib-inline==0.1.6
multidict==6.0.4
nest-asyncio==1.5.8
numpy==1.26.1
packaging==23.2
pandas==2.1.1
parsimonious==0.9.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.11.0
prompt-toolkit==3.0.39
protobuf==4.24.4
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycryptodome==3.19.0
Pygments==2.16.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyunormalize==15.0.0
pyzmq==25.1.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rlp==3.0.0
rpds-py==0.10.6
six==1.16.0
stack-data==0.6.3
toolz==0.12.0
tornado==6.3.3
traitlets==5.12.0
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
wcwidth==0.2.8
web3==6.11.1
websockets==12.0
yarl==1.9.2
